refactor(frontend): split box prop geometry
