+++ engine/examples/editor-playtest/src/box_props/rendering.rs
@@ -1,3 +1,4 @@
+use super::geometry::*;
 use super::*;
 
 #[derive(Copy, Clone)]
